Use Database name to retrieve Vault config
The resource that lives at the path `/database/config/*` in the Vault
API is configuration for a specific database connection, so the name is
more likely to map to a specific Database object.

The `Secret` configured on the Database object is actually a mapping to a
`role` in Vault that provides specific permissions on a database (it's a
secret that happens to be creds for a Vault Db role).