Add additional Vault config
This allows a custom Go template to be passed to form the connection URI
when using Vault, and also setting of the Kubernetes auth endpoint if
the default path was not used (see
https://www.vaultproject.io/docs/auth/kubernetes#authentication).

If using the Vault Agent then the Go template specified in the
`connectionTemplate` parameter will be wrapped in the Consul template on
the annotation read by the agent, otherwise the Go template will be used
directly with the Vault secrets read from the specified db role endpoint
to form the URI used by SchemaHero to connect.

Also adds missing serviceaccounts permission to the role that is bound
to the manager pods watching a db, as this is needed to query the k8s
API for the service account to get the token to use to authn to the
Vault endpoint.

Signed-off-by: Paul Thomson <thomsonp83@gmail.com>