Fix generate tests for plugin architecture
The rqlite driver has been moved to a plugin, so the generate
command returns an error for rqlite schemas. This commit comments
out the rqlite test cases in generate_test.go to fix the failing
CI tests.

The generate command will need to be updated to work with the
plugin architecture in a future PR.

Signed-off-by: Marc Campbell <marc.e.campbell@gmail.com>